# src/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Broers Node.js Technical Test API",
        "version": "1.0.0",
        "description": "Documentación de la API para el test técnico de Broers."
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "components": {
        "securitySchemes": {
            "XTokenAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-token"
            }
        },
        "schemas": {
...
            "HTMLResponse": {
                "type": "string",
                "example": "<!DOCTYPE html>... (contenido HTML)"
            },
            "ErrorInternalResponse": {
                "type": "object",
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": false
                    },
                    "msg": {
                        "type": "string",
                        "example": "Internal server error"
                    }
                }
            },
...
            "InvalidTokenResponse": {
                "type": "object",
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": false
                    },
                    "msg": {
                        "type": "string",
                        "example": "El token es inválido o ha expirado, o la nueva contraseña no es válida"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/auth/reset-password/:token": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Mostrar formulario de recuperación de contraseña",
                "description": "Devuelve el formulario HTML donde el usuario puede restablecer su contraseña utilizando el token proporcionado.",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "description": "Token para restablecer la contraseña",
                        "schema": {
                            "type": "string",
                            "example": "example-token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulario de recuperación de contraseña en HTML",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTMLResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Restablecer la contraseña",
                "description": "Recibe el token y la nueva contraseña del usuario para restablecer la contraseña.",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "description": "Token para restablecer la contraseña",
                        "schema": {
                            "type": "string",
                            "example": "example-token"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "newPassword": {
                                        "type": "string",
                                        "description": "Nueva contraseña del usuario",
                                        "example": "newpassword123"
                                    }
                                },
                                "required": [
                                    "newPassword"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña restablecida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Contraseña restablecida exitosamente"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token inválido o expirado, o la nueva contraseña no es válida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvalidTokenResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error en el servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInternalResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}